Migrate TSLint -> ESLint
Move away from TSLint and over to ESLint. This is mostly done via the schematic provided here: https://github.com/angular-eslint/angular-eslint#migrating-an-angular-cli-project-from-codelyzer-and-tslint

It does, however, appear that the TS config for the e2e project is using a non-standard or perhaps outdated name, and so that had to updated manually.

This includes only the changes needed to get ESLint running without configuration errors, and to remove the deprecated linting packages (TSList, Codelyzer). It does not yet add missing rules or remove the old configs.